requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cloudinary==1.44.1
Django==6.0.3
dj-database-url==3.1.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth>=2.0.0,<4.0.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.2
requests>=2.0.0,<3.0.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
